Clarify function structure instructions
I ran into the problem of creating functions that dealt with collections, since the example uses a collection for testing. It was lost on me that onyx automatically breaks the input into single item segments, so I wrote my functions to try to map over that collection. Ironically, this actually returned a collection of things, so the future Onyx functions were able to actually map, but this broke the test.